Adjust test cases for refactor
